Extract frame to show tracks
