+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_canvas.py
@@ -1,8 +1,7 @@
-import tkinter
 from dataclasses import dataclass
-from typing import Any, Optional
+from typing import Any
 
-from customtkinter import NW, CTkCheckBox, CTkFrame
+from customtkinter import NW, CTkFrame
 from PIL import ImageTk
 
 from OTAnalytics.adapter_ui.abstract_canvas import AbstractCanvas
@@ -34,7 +33,6 @@ class FrameCanvas(AbstractFrameCanvas, CTkFrame):
     def __init__(self, viewmodel: ViewModel, **kwargs: Any) -> None:
         super().__init__(**kwargs)
         self._viewmodel = viewmodel
-        self._show_tracks = tkinter.BooleanVar()
         self._get_widgets()
         self._place_widgets()
         self.introduce_to_viewmodel()
@@ -46,22 +44,11 @@ class FrameCanvas(AbstractFrameCanvas, CTkFrame):
         self.canvas_background = CanvasBackground(
             master=self, viewmodel=self._viewmodel
         )
-        self.button_show_tracks = CTkCheckBox(
-            master=self,
-            text="Show tracks",
-            command=self._update_show_tracks_state,
-            variable=self._show_tracks,
-            onvalue=True,
-            offvalue=False,
-        )
 
     def _place_widgets(self) -> None:
         PADY = 10
         self.canvas_background.grid(
             row=1, column=0, padx=PADX, pady=PADY, sticky=STICKY
-        )
-        self.button_show_tracks.grid(
-            row=0, column=0, padx=PADX, pady=PADY, sticky=STICKY
         )
 
     def update_background(self, image: TrackImage) -> None:
@@ -72,14 +59,6 @@ class FrameCanvas(AbstractFrameCanvas, CTkFrame):
 
     def remove_layer(self, layer: str) -> None:
         self.canvas_background.remove_layer(layer)
-
-    def update_show_tracks(self, value: Optional[bool]) -> None:
-        new_value = value or False
-        self._show_tracks.set(new_value)
-
-    def _update_show_tracks_state(self) -> None:
-        new_value = self._show_tracks.get()
-        self._viewmodel.update_show_tracks_state(new_value)
 
 
 class CanvasBackground(AbstractCanvas):
