+++ OTAnalytics/plugin_ui/customtkinter_gui/gui.py
@@ -7,6 +7,9 @@ from OTAnalytics.plugin_ui.customtkinter_gui.frame_canvas import FrameCanvas
 from OTAnalytics.plugin_ui.customtkinter_gui.frame_configuration import (
     FrameConfiguration,
 )
+from OTAnalytics.plugin_ui.customtkinter_gui.frame_track_plotting import (
+    FrameTrackPlotting,
+)
 from OTAnalytics.plugin_ui.customtkinter_gui.frame_tracks import TracksFrame
 
 
@@ -33,6 +36,10 @@ class OTAnalyticsGui:
         self._app.mainloop()
 
     def _get_widgets(self) -> None:
+        self.frame_track_plotting = FrameTrackPlotting(
+            master=self._app,
+            viewmodel=self._view_model,
+        )
         self.frame_canvas = FrameCanvas(
             master=self._app,
             viewmodel=self._view_model,
@@ -50,8 +57,11 @@ class OTAnalyticsGui:
 
     def _place_widgets(self) -> None:
         PADY = 10
+        self.frame_track_plotting.grid(
+            row=0, column=0, padx=PADX, pady=PADY, sticky=STICKY
+        )
         self.frame_canvas.grid(
-            row=0, column=0, rowspan=3, padx=PADX, pady=PADY, sticky=STICKY
+            row=1, column=0, rowspan=3, padx=PADX, pady=PADY, sticky=STICKY
         )
         self.frame_tracks.grid(row=0, column=1, padx=PADX, pady=PADY, sticky=STICKY)
         self.tabview_configuration.grid(
